restructure MibigBGC and add tests
