+++ src/nplinker/genomics/mibig/__init__.py
@@ -1,8 +1,7 @@
 import logging
+from .mibig_bgc import MibigBGC
 from .mibig_downloader import download_and_extract_mibig_metadata
 from .mibig_metadata import MibigMetadata
 
+
 logging.getLogger(__name__).addHandler(logging.NullHandler())
-
-
-__all__ = {download_and_extract_mibig_metadata}
